# api-docs.json
{"openapi":"3.0.1","info":{"title":"Spring Boot Order Management System REST APIs","description":"Spring Boot Order Management System REST API Documentation","termsOfService":"Terms of service","contact":{"name":"Salah Aldin Dar AlDeek","url":"https://salahaldin2021.github.io/","email":"salahaldin.daraldeek@gmail.com"},"license":{"name":"License of API","url":"https://www.webservicecomp438ordermanagmentapi.com/license"},"version":"1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"Authorization":["global"]}],"paths":{"/api/v1/products/{productId}/stocks/{id}":{"get":{"tags":["Stock"],"summary":"Get specific stock by stock ID of a specific product","operationId":"getStockByProductIdAndStockId","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StockDTO"}}}}}},"put":{"tags":["Stock"],"summary":"Update stock for a specific product","operationId":"updateStockByProductIdAndStockId","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StockDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StockDTO"}}}}}},"delete":{"tags":["Stock"],"summary":"Delete stock by stock ID of a specific product","operationId":"deleteStockByProductIdAndStockId","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/products/{id}":{"get":{"tags":["Products"],"summary":"Get product by ID","operationId":"getProductById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}}}},"put":{"tags":["Products"],"summary":"Update product","operationId":"updateProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}}}},"delete":{"tags":["Products"],"summary":"Delete product by ID","operationId":"deleteProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/customers/{id}":{"get":{"tags":["Customers"],"summary":"Get customer by ID","operationId":"getCustomerById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerDTO"}}}}}},"put":{"tags":["Customers"],"summary":"Update customer","operationId":"updateCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerDTO"}}}}}},"delete":{"tags":["Customers"],"summary":"Delete customer by ID","operationId":"deleteCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/customers/{customerId}/orders/{orderId}":{"get":{"tags":["Orders"],"summary":"Get specific order of a specific customer","operationId":"getOrderByIdAndCustomerId","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderDTO"}}}}}},"put":{"tags":["Orders"],"summary":"Update an order for a specific customer","operationId":"updateOrderForCustomer","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderDTO"}}}}}},"delete":{"tags":["Orders"],"summary":"Delete an order for a specific customer","operationId":"deleteOrderForCustomer","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/customers/{customerId}/orders/{orderId}/products/{productId}":{"get":{"tags":["Product Orders"],"summary":"Get specific product of specific order","operationId":"getProductOfOrder","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"productId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductOrderDTO"}}}}}},"put":{"tags":["Product Orders"],"summary":"Update specific product of specific order","operationId":"updateProductOfOrder","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"productId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductOrderDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductOrderDTO"}}}}}},"delete":{"tags":["Product Orders"],"summary":"Delete specific product of specific order","operationId":"deleteProductOfOrder","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"productId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/products":{"get":{"tags":["Products"],"summary":"Get all products","operationId":"getAllProducts","parameters":[{"name":"pageNo","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"id"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageProductDTO"}}}}}},"post":{"tags":["Products"],"summary":"Create a new product","operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}}}}},"/api/v1/products/{productId}/stocks":{"get":{"tags":["Stock"],"summary":"Get all stocks of a specific product","operationId":"getAllStocksByProductId","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StockDTO"}}}}}}},"post":{"tags":["Stock"],"summary":"Create a new stock entry for a specific product","operationId":"createStockByProductId","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StockDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StockDTO"}}}}}}},"/api/v1/customers":{"get":{"tags":["Customers"],"summary":"Get all customers","operationId":"getAllCustomers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CustomerDTO"}}}}}}},"post":{"tags":["Customers"],"summary":"Create a new customer","operationId":"createCustomer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerDTO"}}}}}}},"/api/v1/customers/{customerId}/orders":{"get":{"tags":["Orders"],"summary":"Get all orders for a specific customer","operationId":"getOrdersByCustomerId","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"pageNo","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"id"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageOrderDTO"}}}}}},"post":{"tags":["Orders"],"summary":"Create an order for a specific customer","operationId":"createOrderForCustomer","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderDTO"}}}}}}},"/api/v1/customers/{customerId}/orders/{orderId}/products":{"get":{"tags":["Product Orders"],"summary":"Get products of specific order","operationId":"getProductsOfOrder","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductOrderDTO"}}}}}}},"post":{"tags":["Product Orders"],"summary":"Post specific product to specific order","operationId":"postProductToOrder","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductOrderDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductOrderDTO"}}}}}}},"/api/v1/auth/signup":{"post":{"tags":["Auth controller exposes siginin and signup REST APIs"],"summary":"Register user","description":"REST API to Signin or Login user to Blog app","operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/signin":{"post":{"tags":["Auth controller exposes siginin and signup REST APIs"],"summary":"Authenticate user","description":"REST API to Register or Signup user to Blog app","operationId":"authenticateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JWTAuthResponse"}}}}}}}},"components":{"schemas":{"StockDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"productId":{"type":"integer","format":"int32"},"quantity":{"type":"integer","format":"int32"},"updateAt":{"type":"string","format":"date-time"}}},"Links":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}},"ProductDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"slug":{"type":"string"},"name":{"type":"string"},"reference":{"type":"string"},"price":{"type":"number"},"vat":{"type":"number"},"stockable":{"type":"boolean"},"_links":{"$ref":"#/components/schemas/Links"}}},"CustomerDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"firstName":{"type":"string"},"lastName":{"type":"string"},"bornAt":{"type":"string","format":"date"},"_links":{"$ref":"#/components/schemas/Links"}}},"OrderDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"customerId":{"type":"integer","format":"int32"},"orderAt":{"type":"string","format":"date-time"},"_links":{"$ref":"#/components/schemas/Links"}}},"ProductOrderDTO":{"type":"object","properties":{"productId":{"type":"integer","format":"int32"},"orderId":{"type":"integer","format":"int32"},"quantity":{"type":"integer","format":"int32"},"price":{"type":"number"},"vat":{"type":"number"},"_links":{"$ref":"#/components/schemas/Links"}}},"SignUpDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"bornAt":{"type":"string","format":"date"}}},"LoginDto":{"type":"object","properties":{"usernameOrEmail":{"type":"string"},"password":{"type":"string"}}},"JWTAuthResponse":{"type":"object","properties":{"token":{"type":"string"}}},"PageProductDTO":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/ProductDTO"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageSize":{"type":"integer","format":"int32"},"pageNumber":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"},"paged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"PageOrderDTO":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/OrderDTO"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"Link":{"type":"object","properties":{"href":{"type":"string"},"hreflang":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"},"deprecation":{"type":"string"},"profile":{"type":"string"},"name":{"type":"string"},"templated":{"type":"boolean"}}}},"securitySchemes":{"Authorization":{"type":"apiKey","name":"Authorization","in":"header"}}}}